fix ird parsing for discs with mastering errors (trailing dots in filenames)
