+++ IrdLibraryClient/IrdFormat/IrdParser.cs
@@ -2,9 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
-using System.Linq;
 using System.Text;
-using DiscUtils.Iso9660;
 using Force.Crc32;
 
 namespace IrdLibraryClient.IrdFormat
@@ -77,23 +75,6 @@ namespace IrdLibraryClient.IrdFormat
                     throw new InvalidDataException($"Corrupted IRD data, expected {result.Crc32:x8}, but was {crc32:x8}");
             }
             return result;
-        }
-    }
-
-    public static class IsoHeaderParser
-    {
-        public static List<string> GetFilenames(this Ird ird)
-        {
-            using (var decompressedStream = new MemoryStream())
-            {
-                using (var compressedStream = new MemoryStream(ird.Header, false))
-                using (var gzip = new GZipStream(compressedStream, CompressionMode.Decompress))
-                    gzip.CopyTo(decompressedStream);
-
-                decompressedStream.Seek(0, SeekOrigin.Begin);
-                var reader = new CDReader(decompressedStream, true, true);
-                return reader.GetFiles(reader.Root.FullName, "*.*", SearchOption.AllDirectories).Distinct().Select(n => n.TrimStart('\\').Replace('\\', '/')).ToList();
-            }
         }
     }
 }
